third_party/WebKit/LayoutTests/platform/android/external/wpt/webvtt/parsing/file-parsing/tests/settings-line-expected.txt

This is a testharness.js-based test.
FAIL settings, lineassert_equals: Failed with cue 9 expected 1e+34 but got 9.999999790214768e+33
Harness: the test ran to completion.
